refactor: 폴더 구조 변경에 따른 내부 경로 수정
