+++ src/app/(route)/community/[community_id]/_component/_style/postDetails.css.ts
