+++ src/app/(route)/community/[community_id]/page.tsx
@@ -1,4 +1,4 @@
-import PostDetails from "../_component/PostDetail";
+import PostDetails from "./_component/PostDetail";
 
 export default function Page({ params }: { params: { community_id: string } }) {
   return (
